move file adapter around a bit and split up the read/write to make it easier to further developer features on the file adapter.
