+++ lib/adapters/file/read.js
@@ -1,18 +1,6 @@
-//
-// Simple Juttle adapter that implements a single file "database".
-//
-// The "filename" option indicates a file to read/write from and it must contain
-// a JSON array of objects.
-//
-// The filters passed in the read parameterrs are executed to filter the points
-// from the file.
-
-/* jshint evil: true */
-
 var _ = require('underscore');
-var Juttle = require('../runtime/index').Juttle;
-var JuttleMoment = require('../moment').JuttleMoment;
-var parsers = require('./parsers');
+var Juttle = require('../../runtime/index').Juttle;
+var parsers = require('../parsers');
 var fs = require('fs');
 var values = require('../runtime/values');
 
@@ -100,66 +88,5 @@ var Read = Juttle.proc.source.extend({
     }
 });
 
-var Write = Juttle.proc.sink.extend({
-    procName: 'write-file',
-    initialize: function(options, params) {
-        var allowed_options = ['file'];
-        var unknown = _.difference(_.keys(options), allowed_options);
-        if (unknown.length > 0) {
-            throw this.compile_error('RT-UNKNOWN-OPTION-ERROR', {
-                proc: 'write file',
-                option: unknown[0]
-            });
-        }
 
-        if (!_.has(options, 'file')) {
-            throw this.compile_error('RT-MISSING-OPTION-ERROR', {
-                proc: 'write file',
-                option: "file"
-            });
-        }
-
-        this.filename = options.file;
-        this.queue = [];
-    },
-
-    process: function(points) {
-        this.logger.debug('process', points);
-        this.queue = this.queue.concat(points);
-    },
-
-    flush: function() {
-        // XXX/demmer this should be asynchronous
-        var data;
-        var points = [];
-        try {
-            data = fs.readFileSync(this.filename, 'utf8');
-            points = JSON.parse(data);
-        } catch (err) {
-            if (err.code !== 'ENOENT') {
-                this.logger.error('error reading file:', err.toString());
-            }
-        }
-        points = points.concat(this.queue);
-        data = JSON.stringify(points, null, 4);
-        fs.writeFileSync(this.filename, data);
-
-        this.queue = [];
-    },
-
-    eof: function() {
-        this.logger.debug('eof');
-        this.flush();
-        this.done();
-    }
-});
-
-function FileAdapter(config) {
-    return {
-        name: 'file',
-        read: Read,
-        write: Write
-    };
-}
-
-module.exports = FileAdapter;
+module.exports = Read;
