+++ lib/adapters/index.js
@@ -1,7 +1,7 @@
 var Juttle = require('../runtime/index').Juttle;
 
 var stochastic = require('./stochastic-adapter');
-var file = require('./file-adapter');
+var file = require('./file');
 var http = require('./http');
 
 // register the built-in adapters
